+++ internal/provider/model_sdwan_service_lan_vpn_interface_ethernet_feature.go
@@ -293,7 +293,12 @@ func (data ServiceLANVPNInterfaceEthernet) toBody(ctx context.Context, currentVe
 			body, _ = sjson.Set(body, path+"intfIpAddress.dynamic.dynamicDhcpDistance.optionType", "variable")
 			body, _ = sjson.Set(body, path+"intfIpAddress.dynamic.dynamicDhcpDistance.value", data.Ipv4DhcpDistanceVariable.ValueString())
 		}
-	} else if !data.Ipv4DhcpDistance.IsNull() {
+	} else if data.Ipv4DhcpDistance.IsNull() {
+		if true && data.Ipv4ConfigurationType.ValueString() == "dynamic" {
+			body, _ = sjson.Set(body, path+"intfIpAddress.dynamic.dynamicDhcpDistance.optionType", "default")
+			body, _ = sjson.Set(body, path+"intfIpAddress.dynamic.dynamicDhcpDistance.value", 1)
+		}
+	} else {
 		if true && data.Ipv4ConfigurationType.ValueString() == "dynamic" {
 			body, _ = sjson.Set(body, path+"intfIpAddress.dynamic.dynamicDhcpDistance.optionType", "global")
 			body, _ = sjson.Set(body, path+"intfIpAddress.dynamic.dynamicDhcpDistance.value", data.Ipv4DhcpDistance.ValueInt64())